Move test intkey client to integration
